requirements.txt
pyats[full]>=24.0
genie>=24.0
pytest>=7.0
netmiko>=4.0
pynetbox>=7.0
jinja2>=3.0
pyyaml>=6.0
